Move pyflakes violations to rules modules
