+++ src/registry.rs
@@ -33,49 +33,49 @@ ruff_macros::define_rule_mapping!(
     W505 => rules::pycodestyle::rules::DocLineTooLong,
     W605 => rules::pycodestyle::rules::InvalidEscapeSequence,
     // pyflakes
-    F401 => violations::UnusedImport,
-    F402 => violations::ImportShadowedByLoopVar,
-    F403 => violations::ImportStarUsed,
-    F404 => violations::LateFutureImport,
-    F405 => violations::ImportStarUsage,
-    F406 => violations::ImportStarNotPermitted,
-    F407 => violations::FutureFeatureNotDefined,
-    F501 => violations::PercentFormatInvalidFormat,
-    F502 => violations::PercentFormatExpectedMapping,
-    F503 => violations::PercentFormatExpectedSequence,
-    F504 => violations::PercentFormatExtraNamedArguments,
-    F505 => violations::PercentFormatMissingArgument,
-    F506 => violations::PercentFormatMixedPositionalAndNamed,
-    F507 => violations::PercentFormatPositionalCountMismatch,
-    F508 => violations::PercentFormatStarRequiresSequence,
-    F509 => violations::PercentFormatUnsupportedFormatCharacter,
-    F521 => violations::StringDotFormatInvalidFormat,
-    F522 => violations::StringDotFormatExtraNamedArguments,
-    F523 => violations::StringDotFormatExtraPositionalArguments,
-    F524 => violations::StringDotFormatMissingArguments,
-    F525 => violations::StringDotFormatMixingAutomatic,
-    F541 => violations::FStringMissingPlaceholders,
-    F601 => violations::MultiValueRepeatedKeyLiteral,
-    F602 => violations::MultiValueRepeatedKeyVariable,
-    F621 => violations::ExpressionsInStarAssignment,
-    F622 => violations::TwoStarredExpressions,
-    F631 => violations::AssertTuple,
-    F632 => violations::IsLiteral,
-    F633 => violations::InvalidPrintSyntax,
-    F634 => violations::IfTuple,
-    F701 => violations::BreakOutsideLoop,
-    F702 => violations::ContinueOutsideLoop,
-    F704 => violations::YieldOutsideFunction,
-    F706 => violations::ReturnOutsideFunction,
-    F707 => violations::DefaultExceptNotLast,
-    F722 => violations::ForwardAnnotationSyntaxError,
-    F811 => violations::RedefinedWhileUnused,
-    F821 => violations::UndefinedName,
-    F822 => violations::UndefinedExport,
-    F823 => violations::UndefinedLocal,
-    F841 => violations::UnusedVariable,
-    F842 => violations::UnusedAnnotation,
-    F901 => violations::RaiseNotImplemented,
+    F401 => rules::pyflakes::rules::UnusedImport,
+    F402 => rules::pyflakes::rules::ImportShadowedByLoopVar,
+    F403 => rules::pyflakes::rules::ImportStarUsed,
+    F404 => rules::pyflakes::rules::LateFutureImport,
+    F405 => rules::pyflakes::rules::ImportStarUsage,
+    F406 => rules::pyflakes::rules::ImportStarNotPermitted,
+    F407 => rules::pyflakes::rules::FutureFeatureNotDefined,
+    F501 => rules::pyflakes::rules::PercentFormatInvalidFormat,
+    F502 => rules::pyflakes::rules::PercentFormatExpectedMapping,
+    F503 => rules::pyflakes::rules::PercentFormatExpectedSequence,
+    F504 => rules::pyflakes::rules::PercentFormatExtraNamedArguments,
+    F505 => rules::pyflakes::rules::PercentFormatMissingArgument,
+    F506 => rules::pyflakes::rules::PercentFormatMixedPositionalAndNamed,
+    F507 => rules::pyflakes::rules::PercentFormatPositionalCountMismatch,
+    F508 => rules::pyflakes::rules::PercentFormatStarRequiresSequence,
+    F509 => rules::pyflakes::rules::PercentFormatUnsupportedFormatCharacter,
+    F521 => rules::pyflakes::rules::StringDotFormatInvalidFormat,
+    F522 => rules::pyflakes::rules::StringDotFormatExtraNamedArguments,
+    F523 => rules::pyflakes::rules::StringDotFormatExtraPositionalArguments,
+    F524 => rules::pyflakes::rules::StringDotFormatMissingArguments,
+    F525 => rules::pyflakes::rules::StringDotFormatMixingAutomatic,
+    F541 => rules::pyflakes::rules::FStringMissingPlaceholders,
+    F601 => rules::pyflakes::rules::MultiValueRepeatedKeyLiteral,
+    F602 => rules::pyflakes::rules::MultiValueRepeatedKeyVariable,
+    F621 => rules::pyflakes::rules::ExpressionsInStarAssignment,
+    F622 => rules::pyflakes::rules::TwoStarredExpressions,
+    F631 => rules::pyflakes::rules::AssertTuple,
+    F632 => rules::pyflakes::rules::IsLiteral,
+    F633 => rules::pyflakes::rules::InvalidPrintSyntax,
+    F634 => rules::pyflakes::rules::IfTuple,
+    F701 => rules::pyflakes::rules::BreakOutsideLoop,
+    F702 => rules::pyflakes::rules::ContinueOutsideLoop,
+    F704 => rules::pyflakes::rules::YieldOutsideFunction,
+    F706 => rules::pyflakes::rules::ReturnOutsideFunction,
+    F707 => rules::pyflakes::rules::DefaultExceptNotLast,
+    F722 => rules::pyflakes::rules::ForwardAnnotationSyntaxError,
+    F811 => rules::pyflakes::rules::RedefinedWhileUnused,
+    F821 => rules::pyflakes::rules::UndefinedName,
+    F822 => rules::pyflakes::rules::UndefinedExport,
+    F823 => rules::pyflakes::rules::UndefinedLocal,
+    F841 => rules::pyflakes::rules::UnusedVariable,
+    F842 => rules::pyflakes::rules::UnusedAnnotation,
+    F901 => rules::pyflakes::rules::RaiseNotImplemented,
     // pylint
     PLE0604 => rules::pylint::rules::InvalidAllObject,
     PLE0605 => rules::pylint::rules::InvalidAllFormat,
