+++ src/rules/pyflakes/rules/assert_tuple.rs
@@ -1,18 +1,29 @@
+use crate::define_violation;
+use ruff_macros::derive_message_formats;
 use rustpython_ast::{Expr, ExprKind, Stmt};
 
 use crate::ast::types::Range;
 use crate::checkers::ast::Checker;
 use crate::registry::Diagnostic;
-use crate::violations;
+use crate::violation::Violation;
+
+define_violation!(
+    pub struct AssertTuple;
+);
+impl Violation for AssertTuple {
+    #[derive_message_formats]
+    fn message(&self) -> String {
+        format!("Assert test is a non-empty tuple, which is always `True`")
+    }
+}
 
 /// F631
 pub fn assert_tuple(checker: &mut Checker, stmt: &Stmt, test: &Expr) {
     if let ExprKind::Tuple { elts, .. } = &test.node {
         if !elts.is_empty() {
-            checker.diagnostics.push(Diagnostic::new(
-                violations::AssertTuple,
-                Range::from_located(stmt),
-            ));
+            checker
+                .diagnostics
+                .push(Diagnostic::new(AssertTuple, Range::from_located(stmt)));
         }
     }
 }
