+++ src/rules/pyflakes/rules/continue_outside_loop.rs
@@ -1,7 +1,20 @@
 use crate::ast::types::Range;
+use crate::define_violation;
 use crate::registry::Diagnostic;
-use crate::violations;
+
+use crate::violation::Violation;
+use ruff_macros::derive_message_formats;
 use rustpython_ast::{Stmt, StmtKind};
+
+define_violation!(
+    pub struct ContinueOutsideLoop;
+);
+impl Violation for ContinueOutsideLoop {
+    #[derive_message_formats]
+    fn message(&self) -> String {
+        format!("`continue` not properly in loop")
+    }
+}
 
 /// F702
 pub fn continue_outside_loop<'a>(
@@ -34,7 +47,7 @@ pub fn continue_outside_loop<'a>(
         None
     } else {
         Some(Diagnostic::new(
-            violations::ContinueOutsideLoop,
+            ContinueOutsideLoop,
             Range::from_located(stmt),
         ))
     }
