+++ src/rules/pyflakes/rules/f_string_missing_placeholders.rs
@@ -1,10 +1,26 @@
+use crate::define_violation;
+use ruff_macros::derive_message_formats;
 use rustpython_ast::{Expr, ExprKind};
 
 use crate::ast::helpers::find_useless_f_strings;
 use crate::checkers::ast::Checker;
 use crate::fix::Fix;
 use crate::registry::Diagnostic;
-use crate::violations;
+use crate::violation::AlwaysAutofixableViolation;
+
+define_violation!(
+    pub struct FStringMissingPlaceholders;
+);
+impl AlwaysAutofixableViolation for FStringMissingPlaceholders {
+    #[derive_message_formats]
+    fn message(&self) -> String {
+        format!("f-string without any placeholders")
+    }
+
+    fn autofix_title(&self) -> String {
+        "Remove extraneous `f` prefix".to_string()
+    }
+}
 
 /// F541
 pub fn f_string_missing_placeholders(expr: &Expr, values: &[Expr], checker: &mut Checker) {
@@ -13,7 +29,7 @@ pub fn f_string_missing_placeholders(expr: &Expr, values: &[Expr], checker: &mut
         .any(|value| matches!(value.node, ExprKind::FormattedValue { .. }))
     {
         for (prefix_range, tok_range) in find_useless_f_strings(expr, checker.locator) {
-            let mut diagnostic = Diagnostic::new(violations::FStringMissingPlaceholders, tok_range);
+            let mut diagnostic = Diagnostic::new(FStringMissingPlaceholders, tok_range);
             if checker.patch(diagnostic.kind.rule()) {
                 diagnostic.amend(Fix::deletion(
                     prefix_range.location,
