+++ src/rules/pyflakes/rules/mod.rs
@@ -1,38 +1,67 @@
-pub use assert_tuple::assert_tuple;
-pub use break_outside_loop::break_outside_loop;
-pub use continue_outside_loop::continue_outside_loop;
-pub use default_except_not_last::default_except_not_last;
-pub use f_string_missing_placeholders::f_string_missing_placeholders;
-pub use if_tuple::if_tuple;
-pub use invalid_literal_comparisons::invalid_literal_comparison;
-pub use invalid_print_syntax::invalid_print_syntax;
-pub use raise_not_implemented::raise_not_implemented;
-pub use repeated_keys::repeated_keys;
-pub use starred_expressions::starred_expressions;
+pub use assert_tuple::{assert_tuple, AssertTuple};
+pub use break_outside_loop::{break_outside_loop, BreakOutsideLoop};
+pub use continue_outside_loop::{continue_outside_loop, ContinueOutsideLoop};
+pub use default_except_not_last::{default_except_not_last, DefaultExceptNotLast};
+pub use f_string_missing_placeholders::{
+    f_string_missing_placeholders, FStringMissingPlaceholders,
+};
+pub use forward_annotation_syntax_error::ForwardAnnotationSyntaxError;
+pub use if_tuple::{if_tuple, IfTuple};
+pub use imports::{
+    future_feature_not_defined, FutureFeatureNotDefined, ImportShadowedByLoopVar,
+    ImportStarNotPermitted, ImportStarUsage, ImportStarUsed, LateFutureImport, UnusedImport,
+};
+pub use invalid_literal_comparisons::{invalid_literal_comparison, IsLiteral};
+pub use invalid_print_syntax::{invalid_print_syntax, InvalidPrintSyntax};
+pub use raise_not_implemented::{raise_not_implemented, RaiseNotImplemented};
+pub use redefined_while_unused::RedefinedWhileUnused;
+pub use repeated_keys::{
+    repeated_keys, MultiValueRepeatedKeyLiteral, MultiValueRepeatedKeyVariable,
+};
+pub use return_outside_function::{return_outside_function, ReturnOutsideFunction};
+pub use starred_expressions::{
+    starred_expressions, ExpressionsInStarAssignment, TwoStarredExpressions,
+};
 pub(crate) use strings::{
     percent_format_expected_mapping, percent_format_expected_sequence,
     percent_format_extra_named_arguments, percent_format_missing_arguments,
     percent_format_mixed_positional_and_named, percent_format_positional_count_mismatch,
     percent_format_star_requires_sequence, string_dot_format_extra_named_arguments,
     string_dot_format_extra_positional_arguments, string_dot_format_missing_argument,
-    string_dot_format_mixing_automatic,
+    string_dot_format_mixing_automatic, PercentFormatExpectedMapping,
+    PercentFormatExpectedSequence, PercentFormatExtraNamedArguments, PercentFormatInvalidFormat,
+    PercentFormatMissingArgument, PercentFormatMixedPositionalAndNamed,
+    PercentFormatPositionalCountMismatch, PercentFormatStarRequiresSequence,
+    PercentFormatUnsupportedFormatCharacter, StringDotFormatExtraNamedArguments,
+    StringDotFormatExtraPositionalArguments, StringDotFormatInvalidFormat,
+    StringDotFormatMissingArguments, StringDotFormatMixingAutomatic,
 };
-pub use undefined_local::undefined_local;
-pub use unused_annotation::unused_annotation;
-pub use unused_variable::unused_variable;
+pub use undefined_export::UndefinedExport;
+pub use undefined_local::{undefined_local, UndefinedLocal};
+pub use undefined_name::UndefinedName;
+pub use unused_annotation::{unused_annotation, UnusedAnnotation};
+pub use unused_variable::{unused_variable, UnusedVariable};
+pub use yield_outside_function::{yield_outside_function, YieldOutsideFunction};
 
 mod assert_tuple;
 mod break_outside_loop;
 mod continue_outside_loop;
 mod default_except_not_last;
 mod f_string_missing_placeholders;
+mod forward_annotation_syntax_error;
 mod if_tuple;
+mod imports;
 mod invalid_literal_comparisons;
 mod invalid_print_syntax;
 mod raise_not_implemented;
+mod redefined_while_unused;
 mod repeated_keys;
+mod return_outside_function;
 mod starred_expressions;
 mod strings;
+mod undefined_export;
 mod undefined_local;
+mod undefined_name;
 mod unused_annotation;
 mod unused_variable;
+mod yield_outside_function;
