+++ src/rules/pyflakes/rules/raise_not_implemented.rs
@@ -1,10 +1,26 @@
+use crate::define_violation;
+use ruff_macros::derive_message_formats;
 use rustpython_ast::{Expr, ExprKind};
 
 use crate::ast::types::Range;
 use crate::checkers::ast::Checker;
 use crate::fix::Fix;
 use crate::registry::Diagnostic;
-use crate::violations;
+use crate::violation::AlwaysAutofixableViolation;
+
+define_violation!(
+    pub struct RaiseNotImplemented;
+);
+impl AlwaysAutofixableViolation for RaiseNotImplemented {
+    #[derive_message_formats]
+    fn message(&self) -> String {
+        format!("`raise NotImplemented` should be `raise NotImplementedError`")
+    }
+
+    fn autofix_title(&self) -> String {
+        "Use `raise NotImplementedError`".to_string()
+    }
+}
 
 fn match_not_implemented(expr: &Expr) -> Option<&Expr> {
     match &expr.node {
@@ -30,8 +46,7 @@ pub fn raise_not_implemented(checker: &mut Checker, expr: &Expr) {
     let Some(expr) = match_not_implemented(expr) else {
         return;
     };
-    let mut diagnostic =
-        Diagnostic::new(violations::RaiseNotImplemented, Range::from_located(expr));
+    let mut diagnostic = Diagnostic::new(RaiseNotImplemented, Range::from_located(expr));
     if checker.patch(diagnostic.kind.rule()) {
         diagnostic.amend(Fix::replacement(
             "NotImplementedError".to_string(),
