+++ src/rules/pyflakes/rules/repeated_keys.rs
@@ -9,7 +9,63 @@ use crate::ast::types::Range;
 use crate::checkers::ast::Checker;
 use crate::fix::Fix;
 use crate::registry::{Diagnostic, Rule};
-use crate::violations;
+use crate::{define_violation, AutofixKind};
+
+use crate::violation::{Availability, Violation};
+use ruff_macros::derive_message_formats;
+
+define_violation!(
+    pub struct MultiValueRepeatedKeyLiteral {
+        pub name: String,
+        pub repeated_value: bool,
+    }
+);
+impl Violation for MultiValueRepeatedKeyLiteral {
+    const AUTOFIX: Option<AutofixKind> = Some(AutofixKind::new(Availability::Always));
+
+    #[derive_message_formats]
+    fn message(&self) -> String {
+        let MultiValueRepeatedKeyLiteral { name, .. } = self;
+        format!("Dictionary key literal `{name}` repeated")
+    }
+
+    fn autofix_title_formatter(&self) -> Option<fn(&Self) -> String> {
+        let MultiValueRepeatedKeyLiteral { repeated_value, .. } = self;
+        if *repeated_value {
+            Some(|MultiValueRepeatedKeyLiteral { name, .. }| {
+                format!("Remove repeated key literal `{name}`")
+            })
+        } else {
+            None
+        }
+    }
+}
+define_violation!(
+    pub struct MultiValueRepeatedKeyVariable {
+        pub name: String,
+        pub repeated_value: bool,
+    }
+);
+impl Violation for MultiValueRepeatedKeyVariable {
+    const AUTOFIX: Option<AutofixKind> = Some(AutofixKind::new(Availability::Always));
+
+    #[derive_message_formats]
+    fn message(&self) -> String {
+        let MultiValueRepeatedKeyVariable { name, .. } = self;
+        format!("Dictionary key `{name}` repeated")
+    }
+
+    fn autofix_title_formatter(&self) -> Option<fn(&Self) -> String> {
+        let MultiValueRepeatedKeyVariable { repeated_value, .. } = self;
+        if *repeated_value {
+            Some(|MultiValueRepeatedKeyVariable { name, .. }| {
+                format!("Remove repeated key `{name}`")
+            })
+        } else {
+            None
+        }
+    }
+}
 
 #[derive(Debug, Eq, PartialEq, Hash)]
 enum DictionaryKey<'a> {
@@ -48,7 +104,7 @@ pub fn repeated_keys(checker: &mut Checker, keys: &[Option<Expr>], values: &[Exp
                             let comparable_value: ComparableExpr = (&values[i]).into();
                             let is_duplicate_value = seen_values.contains(&comparable_value);
                             let mut diagnostic = Diagnostic::new(
-                                violations::MultiValueRepeatedKeyLiteral {
+                                MultiValueRepeatedKeyLiteral {
                                     name: unparse_expr(key, checker.stylist),
                                     repeated_value: is_duplicate_value,
                                 },
@@ -76,7 +132,7 @@ pub fn repeated_keys(checker: &mut Checker, keys: &[Option<Expr>], values: &[Exp
                             let comparable_value: ComparableExpr = (&values[i]).into();
                             let is_duplicate_value = seen_values.contains(&comparable_value);
                             let mut diagnostic = Diagnostic::new(
-                                violations::MultiValueRepeatedKeyVariable {
+                                MultiValueRepeatedKeyVariable {
                                     name: dict_key.to_string(),
                                     repeated_value: is_duplicate_value,
                                 },
