+++ src/rules/pyflakes/rules/undefined_local.rs
@@ -1,9 +1,25 @@
 use crate::ast::types::{Binding, Scope, ScopeKind};
 use crate::registry::Diagnostic;
 
-use crate::violations;
+use crate::define_violation;
 
+use ruff_macros::derive_message_formats;
+
+use crate::violation::Violation;
 use std::string::ToString;
+
+define_violation!(
+    pub struct UndefinedLocal {
+        pub name: String,
+    }
+);
+impl Violation for UndefinedLocal {
+    #[derive_message_formats]
+    fn message(&self) -> String {
+        let UndefinedLocal { name } = self;
+        format!("Local variable `{name}` referenced before assignment")
+    }
+}
 
 /// F821
 pub fn undefined_local(name: &str, scopes: &[&Scope], bindings: &[Binding]) -> Option<Diagnostic> {
@@ -15,7 +31,7 @@ pub fn undefined_local(name: &str, scopes: &[&Scope], bindings: &[Binding]) -> O
                     if let Some((scope_id, location)) = binding.runtime_usage {
                         if scope_id == current.id {
                             return Some(Diagnostic::new(
-                                violations::UndefinedLocal {
+                                UndefinedLocal {
                                     name: name.to_string(),
                                 },
                                 location,
