+++ src/rules/pyflakes/rules/unused_annotation.rs
@@ -1,7 +1,23 @@
 use crate::ast::types::BindingKind;
 use crate::checkers::ast::Checker;
+use crate::define_violation;
 use crate::registry::Diagnostic;
-use crate::violations;
+
+use crate::violation::Violation;
+use ruff_macros::derive_message_formats;
+
+define_violation!(
+    pub struct UnusedAnnotation {
+        pub name: String,
+    }
+);
+impl Violation for UnusedAnnotation {
+    #[derive_message_formats]
+    fn message(&self) -> String {
+        let UnusedAnnotation { name } = self;
+        format!("Local variable `{name}` is annotated but never used")
+    }
+}
 
 /// F842
 pub fn unused_annotation(checker: &mut Checker, scope: usize) {
@@ -16,7 +32,7 @@ pub fn unused_annotation(checker: &mut Checker, scope: usize) {
             && !checker.settings.dummy_variable_rgx.is_match(name)
         {
             checker.diagnostics.push(Diagnostic::new(
-                violations::UnusedAnnotation {
+                UnusedAnnotation {
                     name: (*name).to_string(),
                 },
                 binding.range,
