+++ src/violations.rs
@@ -62,656 +62,6 @@ impl Violation for SyntaxError {
     }
 }
 
-// pyflakes
-
-define_violation!(
-    pub struct UnusedImport {
-        pub name: String,
-        pub ignore_init: bool,
-        pub multiple: bool,
-    }
-);
-fn fmt_unused_import_autofix_msg(unused_import: &UnusedImport) -> String {
-    let UnusedImport { name, multiple, .. } = unused_import;
-    if *multiple {
-        "Remove unused import".to_string()
-    } else {
-        format!("Remove unused import: `{name}`")
-    }
-}
-impl Violation for UnusedImport {
-    const AUTOFIX: Option<AutofixKind> = Some(AutofixKind::new(Availability::Always));
-
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let UnusedImport {
-            name, ignore_init, ..
-        } = self;
-        if *ignore_init {
-            format!(
-                "`{name}` imported but unused; consider adding to `__all__` or using a redundant \
-                 alias"
-            )
-        } else {
-            format!("`{name}` imported but unused")
-        }
-    }
-
-    fn autofix_title_formatter(&self) -> Option<fn(&Self) -> String> {
-        let UnusedImport { ignore_init, .. } = self;
-        if *ignore_init {
-            None
-        } else {
-            Some(fmt_unused_import_autofix_msg)
-        }
-    }
-}
-
-define_violation!(
-    pub struct ImportShadowedByLoopVar {
-        pub name: String,
-        pub line: usize,
-    }
-);
-impl Violation for ImportShadowedByLoopVar {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let ImportShadowedByLoopVar { name, line } = self;
-        format!("Import `{name}` from line {line} shadowed by loop variable")
-    }
-}
-
-define_violation!(
-    pub struct ImportStarUsed {
-        pub name: String,
-    }
-);
-impl Violation for ImportStarUsed {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let ImportStarUsed { name } = self;
-        format!("`from {name} import *` used; unable to detect undefined names")
-    }
-}
-
-define_violation!(
-    pub struct LateFutureImport;
-);
-impl Violation for LateFutureImport {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("`from __future__` imports must occur at the beginning of the file")
-    }
-}
-
-define_violation!(
-    pub struct ImportStarUsage {
-        pub name: String,
-        pub sources: Vec<String>,
-    }
-);
-impl Violation for ImportStarUsage {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let ImportStarUsage { name, sources } = self;
-        let sources = sources
-            .iter()
-            .map(|source| format!("`{source}`"))
-            .join(", ");
-        format!("`{name}` may be undefined, or defined from star imports: {sources}")
-    }
-}
-
-define_violation!(
-    pub struct ImportStarNotPermitted {
-        pub name: String,
-    }
-);
-impl Violation for ImportStarNotPermitted {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let ImportStarNotPermitted { name } = self;
-        format!("`from {name} import *` only allowed at module level")
-    }
-}
-
-define_violation!(
-    pub struct FutureFeatureNotDefined {
-        pub name: String,
-    }
-);
-impl Violation for FutureFeatureNotDefined {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let FutureFeatureNotDefined { name } = self;
-        format!("Future feature `{name}` is not defined")
-    }
-}
-
-define_violation!(
-    pub struct PercentFormatInvalidFormat {
-        pub message: String,
-    }
-);
-impl Violation for PercentFormatInvalidFormat {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let PercentFormatInvalidFormat { message } = self;
-        format!("`%`-format string has invalid format string: {message}")
-    }
-}
-
-define_violation!(
-    pub struct PercentFormatExpectedMapping;
-);
-impl Violation for PercentFormatExpectedMapping {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("`%`-format string expected mapping but got sequence")
-    }
-}
-
-define_violation!(
-    pub struct PercentFormatExpectedSequence;
-);
-impl Violation for PercentFormatExpectedSequence {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("`%`-format string expected sequence but got mapping")
-    }
-}
-
-define_violation!(
-    pub struct PercentFormatExtraNamedArguments {
-        pub missing: Vec<String>,
-    }
-);
-impl AlwaysAutofixableViolation for PercentFormatExtraNamedArguments {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let PercentFormatExtraNamedArguments { missing } = self;
-        let message = missing.join(", ");
-        format!("`%`-format string has unused named argument(s): {message}")
-    }
-
-    fn autofix_title(&self) -> String {
-        let PercentFormatExtraNamedArguments { missing } = self;
-        let message = missing.join(", ");
-        format!("Remove extra named arguments: {message}")
-    }
-}
-
-define_violation!(
-    pub struct PercentFormatMissingArgument {
-        pub missing: Vec<String>,
-    }
-);
-impl Violation for PercentFormatMissingArgument {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let PercentFormatMissingArgument { missing } = self;
-        let message = missing.join(", ");
-        format!("`%`-format string is missing argument(s) for placeholder(s): {message}")
-    }
-}
-
-define_violation!(
-    pub struct PercentFormatMixedPositionalAndNamed;
-);
-impl Violation for PercentFormatMixedPositionalAndNamed {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("`%`-format string has mixed positional and named placeholders")
-    }
-}
-
-define_violation!(
-    pub struct PercentFormatPositionalCountMismatch {
-        pub wanted: usize,
-        pub got: usize,
-    }
-);
-impl Violation for PercentFormatPositionalCountMismatch {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let PercentFormatPositionalCountMismatch { wanted, got } = self;
-        format!("`%`-format string has {wanted} placeholder(s) but {got} substitution(s)")
-    }
-}
-
-define_violation!(
-    pub struct PercentFormatStarRequiresSequence;
-);
-impl Violation for PercentFormatStarRequiresSequence {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("`%`-format string `*` specifier requires sequence")
-    }
-}
-
-define_violation!(
-    pub struct PercentFormatUnsupportedFormatCharacter {
-        pub char: char,
-    }
-);
-impl Violation for PercentFormatUnsupportedFormatCharacter {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let PercentFormatUnsupportedFormatCharacter { char } = self;
-        format!("`%`-format string has unsupported format character '{char}'")
-    }
-}
-
-define_violation!(
-    pub struct StringDotFormatInvalidFormat {
-        pub message: String,
-    }
-);
-impl Violation for StringDotFormatInvalidFormat {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let StringDotFormatInvalidFormat { message } = self;
-        format!("`.format` call has invalid format string: {message}")
-    }
-}
-
-define_violation!(
-    pub struct StringDotFormatExtraNamedArguments {
-        pub missing: Vec<String>,
-    }
-);
-impl AlwaysAutofixableViolation for StringDotFormatExtraNamedArguments {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let StringDotFormatExtraNamedArguments { missing } = self;
-        let message = missing.join(", ");
-        format!("`.format` call has unused named argument(s): {message}")
-    }
-
-    fn autofix_title(&self) -> String {
-        let StringDotFormatExtraNamedArguments { missing } = self;
-        let message = missing.join(", ");
-        format!("Remove extra named arguments: {message}")
-    }
-}
-
-define_violation!(
-    pub struct StringDotFormatExtraPositionalArguments {
-        pub missing: Vec<String>,
-    }
-);
-impl Violation for StringDotFormatExtraPositionalArguments {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let StringDotFormatExtraPositionalArguments { missing } = self;
-        let message = missing.join(", ");
-        format!("`.format` call has unused arguments at position(s): {message}")
-    }
-}
-
-define_violation!(
-    pub struct StringDotFormatMissingArguments {
-        pub missing: Vec<String>,
-    }
-);
-impl Violation for StringDotFormatMissingArguments {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let StringDotFormatMissingArguments { missing } = self;
-        let message = missing.join(", ");
-        format!("`.format` call is missing argument(s) for placeholder(s): {message}")
-    }
-}
-
-define_violation!(
-    pub struct StringDotFormatMixingAutomatic;
-);
-impl Violation for StringDotFormatMixingAutomatic {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("`.format` string mixes automatic and manual numbering")
-    }
-}
-
-define_violation!(
-    pub struct FStringMissingPlaceholders;
-);
-impl AlwaysAutofixableViolation for FStringMissingPlaceholders {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("f-string without any placeholders")
-    }
-
-    fn autofix_title(&self) -> String {
-        "Remove extraneous `f` prefix".to_string()
-    }
-}
-
-define_violation!(
-    pub struct MultiValueRepeatedKeyLiteral {
-        pub name: String,
-        pub repeated_value: bool,
-    }
-);
-impl Violation for MultiValueRepeatedKeyLiteral {
-    const AUTOFIX: Option<AutofixKind> = Some(AutofixKind::new(Availability::Always));
-
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let MultiValueRepeatedKeyLiteral { name, .. } = self;
-        format!("Dictionary key literal `{name}` repeated")
-    }
-
-    fn autofix_title_formatter(&self) -> Option<fn(&Self) -> String> {
-        let MultiValueRepeatedKeyLiteral { repeated_value, .. } = self;
-        if *repeated_value {
-            Some(|MultiValueRepeatedKeyLiteral { name, .. }| {
-                format!("Remove repeated key literal `{name}`")
-            })
-        } else {
-            None
-        }
-    }
-}
-
-define_violation!(
-    pub struct MultiValueRepeatedKeyVariable {
-        pub name: String,
-        pub repeated_value: bool,
-    }
-);
-impl Violation for MultiValueRepeatedKeyVariable {
-    const AUTOFIX: Option<AutofixKind> = Some(AutofixKind::new(Availability::Always));
-
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let MultiValueRepeatedKeyVariable { name, .. } = self;
-        format!("Dictionary key `{name}` repeated")
-    }
-
-    fn autofix_title_formatter(&self) -> Option<fn(&Self) -> String> {
-        let MultiValueRepeatedKeyVariable { repeated_value, .. } = self;
-        if *repeated_value {
-            Some(|MultiValueRepeatedKeyVariable { name, .. }| {
-                format!("Remove repeated key `{name}`")
-            })
-        } else {
-            None
-        }
-    }
-}
-
-define_violation!(
-    pub struct ExpressionsInStarAssignment;
-);
-impl Violation for ExpressionsInStarAssignment {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("Too many expressions in star-unpacking assignment")
-    }
-}
-
-define_violation!(
-    pub struct TwoStarredExpressions;
-);
-impl Violation for TwoStarredExpressions {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("Two starred expressions in assignment")
-    }
-}
-
-define_violation!(
-    pub struct AssertTuple;
-);
-impl Violation for AssertTuple {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("Assert test is a non-empty tuple, which is always `True`")
-    }
-}
-
-#[derive(Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub enum IsCmpop {
-    Is,
-    IsNot,
-}
-
-impl From<&Cmpop> for IsCmpop {
-    fn from(cmpop: &Cmpop) -> Self {
-        match cmpop {
-            Cmpop::Is => IsCmpop::Is,
-            Cmpop::IsNot => IsCmpop::IsNot,
-            _ => unreachable!("Expected Cmpop::Is | Cmpop::IsNot"),
-        }
-    }
-}
-
-define_violation!(
-    pub struct IsLiteral {
-        pub cmpop: IsCmpop,
-    }
-);
-impl AlwaysAutofixableViolation for IsLiteral {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let IsLiteral { cmpop } = self;
-        match cmpop {
-            IsCmpop::Is => format!("Use `==` to compare constant literals"),
-            IsCmpop::IsNot => format!("Use `!=` to compare constant literals"),
-        }
-    }
-
-    fn autofix_title(&self) -> String {
-        let IsLiteral { cmpop } = self;
-        match cmpop {
-            IsCmpop::Is => "Replace `is` with `==`".to_string(),
-            IsCmpop::IsNot => "Replace `is not` with `!=`".to_string(),
-        }
-    }
-}
-
-define_violation!(
-    pub struct InvalidPrintSyntax;
-);
-impl Violation for InvalidPrintSyntax {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("Use of `>>` is invalid with `print` function")
-    }
-}
-
-define_violation!(
-    pub struct IfTuple;
-);
-impl Violation for IfTuple {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("If test is a tuple, which is always `True`")
-    }
-}
-
-define_violation!(
-    pub struct BreakOutsideLoop;
-);
-impl Violation for BreakOutsideLoop {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("`break` outside loop")
-    }
-}
-
-define_violation!(
-    pub struct ContinueOutsideLoop;
-);
-impl Violation for ContinueOutsideLoop {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("`continue` not properly in loop")
-    }
-}
-
-#[derive(Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub enum DeferralKeyword {
-    Yield,
-    YieldFrom,
-    Await,
-}
-
-impl fmt::Display for DeferralKeyword {
-    fn fmt(&self, fmt: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            DeferralKeyword::Yield => fmt.write_str("yield"),
-            DeferralKeyword::YieldFrom => fmt.write_str("yield from"),
-            DeferralKeyword::Await => fmt.write_str("await"),
-        }
-    }
-}
-
-define_violation!(
-    pub struct YieldOutsideFunction {
-        pub keyword: DeferralKeyword,
-    }
-);
-impl Violation for YieldOutsideFunction {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let YieldOutsideFunction { keyword } = self;
-        format!("`{keyword}` statement outside of a function")
-    }
-}
-
-define_violation!(
-    pub struct ReturnOutsideFunction;
-);
-impl Violation for ReturnOutsideFunction {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("`return` statement outside of a function/method")
-    }
-}
-
-define_violation!(
-    pub struct DefaultExceptNotLast;
-);
-impl Violation for DefaultExceptNotLast {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("An `except` block as not the last exception handler")
-    }
-}
-
-define_violation!(
-    pub struct ForwardAnnotationSyntaxError {
-        pub body: String,
-    }
-);
-impl Violation for ForwardAnnotationSyntaxError {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let ForwardAnnotationSyntaxError { body } = self;
-        format!("Syntax error in forward annotation: `{body}`")
-    }
-}
-
-define_violation!(
-    pub struct RedefinedWhileUnused {
-        pub name: String,
-        pub line: usize,
-    }
-);
-impl Violation for RedefinedWhileUnused {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let RedefinedWhileUnused { name, line } = self;
-        format!("Redefinition of unused `{name}` from line {line}")
-    }
-}
-
-define_violation!(
-    pub struct UndefinedName {
-        pub name: String,
-    }
-);
-impl Violation for UndefinedName {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let UndefinedName { name } = self;
-        format!("Undefined name `{name}`")
-    }
-}
-
-define_violation!(
-    pub struct UndefinedExport {
-        pub name: String,
-    }
-);
-impl Violation for UndefinedExport {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let UndefinedExport { name } = self;
-        format!("Undefined name `{name}` in `__all__`")
-    }
-}
-
-define_violation!(
-    pub struct UndefinedLocal {
-        pub name: String,
-    }
-);
-impl Violation for UndefinedLocal {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let UndefinedLocal { name } = self;
-        format!("Local variable `{name}` referenced before assignment")
-    }
-}
-
-define_violation!(
-    pub struct UnusedVariable {
-        pub name: String,
-    }
-);
-impl AlwaysAutofixableViolation for UnusedVariable {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let UnusedVariable { name } = self;
-        format!("Local variable `{name}` is assigned to but never used")
-    }
-
-    fn autofix_title(&self) -> String {
-        let UnusedVariable { name } = self;
-        format!("Remove assignment to unused variable `{name}`")
-    }
-}
-
-define_violation!(
-    pub struct UnusedAnnotation {
-        pub name: String,
-    }
-);
-impl Violation for UnusedAnnotation {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let UnusedAnnotation { name } = self;
-        format!("Local variable `{name}` is annotated but never used")
-    }
-}
-
-define_violation!(
-    pub struct RaiseNotImplemented;
-);
-impl AlwaysAutofixableViolation for RaiseNotImplemented {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("`raise NotImplemented` should be `raise NotImplementedError`")
-    }
-
-    fn autofix_title(&self) -> String {
-        "Use `raise NotImplementedError`".to_string()
-    }
-}
-
 // pylint
 
 define_violation!(
